+++ hefquin-access/src/test/java/se/liu/ida/hefquin/federation/access/impl/FederationAccessManagerWithChronicleMapCacheTest.java
@@ -790,7 +790,7 @@ public class FederationAccessManagerWithChronicleMapCacheTest extends Federation
 
 			// Wrap in a list and return the response
 			List<SolutionMapping> mockResult = Collections.singletonList( solMap );
-			return new SolMapsResponseImpl( mockResult, fm, req, new Date() );
+			return new SolMapsResponseImpl( mockResult, new Date() );
 		}
 	}
 
@@ -820,8 +820,6 @@ public class FederationAccessManagerWithChronicleMapCacheTest extends Federation
 			final TPFResponse r = new TPFResponseImpl( Collections.emptyList(),
 			                                           Collections.emptyList(),
 			                                           null,
-			                                           fm,
-			                                           req,
 			                                           new Date() ) {
 				final int cardinalityEstimate = cardinality;
 
@@ -840,8 +838,6 @@ public class FederationAccessManagerWithChronicleMapCacheTest extends Federation
 			final TPFResponse r = new TPFResponseImpl( Collections.emptyList(),
 			                                           Collections.emptyList(),
 			                                           null,
-			                                           fm,
-			                                           req,
 			                                           new Date() ) {
 				final int cardinalityEstimate = cardinality;
 
@@ -881,8 +877,6 @@ public class FederationAccessManagerWithChronicleMapCacheTest extends Federation
 			final TPFResponse r = new TPFResponseImpl( Collections.emptyList(),
 			                                           Collections.emptyList(),
 			                                           null,
-			                                           fm,
-			                                           req,
 			                                           new Date() ) {
 				final int cardinalityEstimate = cardinality;